Adding score percent in response and answe

+++ src/main/java/com/facilio/qa/context/questions/BaseMCQContext.java
@@ -21,7 +21,7 @@ public abstract class BaseMCQContext extends QuestionContext {
     @AllArgsConstructor
     public static class OptionSummary {
         private Long option;
-        private Double percent;
+        private Float percent;
         private Integer count;
     }
 }
